requirements.txt
langchain>=0.2,<0.4
langchain-community>=0.2,<0.4
langchain-google-genai>=2.0.10,<3.0.0
langchain-ollama>=0.1.0
google-generativeai>=0.8.5
requests>=2.31.0
pandas==2.1.4
numpy==1.26.2
scipy>=1.11.0
matplotlib==3.8.2
seaborn==0.13.0
plotly>=5.18.0
python-dotenv==1.0.0
tabulate==0.9.0
streamlit==1.29.0
openpyxl>=3.1.2
fastapi>=0.111.0
uvicorn>=0.30.0
requests>=2.31.0
python-multipart>=0.0.6

